small fixes

+++ src/engine/groupby.cpp
@@ -7,7 +7,7 @@ K* distinct(K* list) {
     // 4096 is a very fast capacity from benchmarks
     IntHashMap map(4096);
 
-    for (int64_t i = 0; i < list->n; ++i) {
+    for (int32_t i = 0; i < list->n; ++i) {
         map.get(list->I[i]);
     }
